refactored hole into it's own file and styled some more
