+++ lib/ball.js
@@ -6,25 +6,17 @@ function Ball(x, y) {
   this.v = 7;
 }
 
-// Ball.prototype.update = function() {
-//   this.bodies = this.bodies.filter(notColliding)
-// };
 Ball.prototype.draw = function(ctx) {
   ctx.strokeStyle = '#000000';
+  ctx.fillStyle = '#ffffff';
+  ctx.lineWidth = 4;
   ctx.beginPath();
   ctx.arc(this.x, this.y, this.radius, 0, 2*Math.PI);
   ctx.stroke();
+  ctx.fill();
 
   return this;
 };
-
-
-// Ball.prototype.move = function(r) {
-//   if (this.keyboarder.isDown(this.keyboarder.KEYS.SPACE) && this.v > 0) {
-//     this.x = this.x + this.v/100;
-//     this.y = this.y + this.v/100;
-//     this.v = this.v - r;
-//     console.log(this.v);
 
 Ball.prototype.move = function(direction, resistance = 0.1) {
   let xMovement = Math.cos(direction);
@@ -38,7 +30,7 @@ Ball.prototype.move = function(direction, resistance = 0.1) {
 
 function isColliding(b1,b2) {
   return false;
-};
+}
