+++ lib/index.js
@@ -1,6 +1,7 @@
 const Spinner = require('./spinner');
+const PowerBar = require('./power-bar');
 const Ball = require('./ball');
-const PowerBar = require('./power-bar')
+const Hole = require('./hole');
 
 const canvas = document.getElementById("game-canvas");
 const ctx = canvas.getContext("2d");
@@ -26,7 +27,6 @@ requestAnimationFrame(function gameLoop() {
   powerBar.draw(ctx);
   powerBar.fill(ctx);
   ball.draw(ctx);
-
   hole.draw(ctx);
 
   let direction = Math.PI/2;
@@ -35,22 +35,4 @@ requestAnimationFrame(function gameLoop() {
 
 
   requestAnimationFrame(gameLoop);
-
-
-
-
 });
-
-function Hole(x, y) {
-  this.x = x;
-  this.y = y;
-  this.radius = 5;
-}
-Hole.prototype.draw = function(ctx) {
-  ctx.strokeStyle = '#000000';
-  ctx.beginPath();
-  ctx.arc(this.x, this.y, this.radius, 0, 2*Math.PI);
-  ctx.stroke();
-
-  return this;
-};
