+++ lib/power-bar.js
@@ -7,15 +7,21 @@ function PowerBar() {
 }
 
 PowerBar.prototype.draw = function(ctx) {
-  ctx.strokeStyle = 'black';
+  ctx.strokeStyle = '#ffffff';
+  ctx.fillStyle = '#ffffff';
+  ctx.lineWidth = 2;
   ctx.strokeRect(this.x,this.y,this.width,this.height);
+  ctx.fillRect(this.x,this.y,this.width,this.height);
 
 };
 
 PowerBar.prototype.fill = function(ctx) {
   if (this.size <= 100) {
-    ctx.fillStyle = 'green';
-    ctx.fillRect(this.x, this.y+this.height, this.width, -this.size);
+    ctx.strokeStyle = '#ffffff';
+    ctx.fillStyle = '#ff0000';
+    ctx.lineWidth = 2;
+    ctx.strokeRect(this.x, this.y + this.height, this.width, -this.size);
+    ctx.fillRect(this.x, this.y + this.height, this.width, -this.size);
     this.size ++;
   } else {
     this.size = 1;
