+++ lib/spinner.js
@@ -14,6 +14,9 @@ Spinner.prototype.draw = function(ctx) {
   ctx.beginPath();
   ctx.arc(this.x, this.y, this.radius, this.sAngle, this.eAngle);
   ctx.stroke();
+  ctx.closePath();
+  ctx.fillStyle = '#ffffff';
+  ctx.fill();
 
   return this;
 };
